+++ lib/src/utils/utils.dart
@@ -7,13 +7,14 @@ class UtilTable {
     required BuildContext context,
     required Map<String, dynamic> data,
     required DatatableHeader header,
+    required Duration timeToSubtract,
     Function(Map<String, dynamic> vaue, DatatableHeader header)? onChanged,
     TextEditingController? controller,
   }) async {
     final DateTime? picked = await showDatePicker(
       context: context,
       initialDate: DateTime.now(), // Refer step 1
-      firstDate: DateTime.now(),
+      firstDate: DateTime.now().subtract(timeToSubtract),
       lastDate: DateTime.now().add(const Duration(hours: 43800)),
     );