+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 0.0.4
 
 - add format list
+
+## 0.0.5
+
+- add param `timeToSubtract`, is the time that is restored to the start date in the date inputs
+
+## 0.0.6
+
+- add `isExpanded` in DropdownButtonFormField